+++ src/noahtheduke/splint.clj
@@ -114,6 +114,7 @@
    noahtheduke.splint.rules.style.prefer-condp
    noahtheduke.splint.rules.style.prefer-for-with-literals
    noahtheduke.splint.rules.style.prefer-vary-meta
+   noahtheduke.splint.rules.style.prefixed-libspecs
    noahtheduke.splint.rules.style.reduce-str
    noahtheduke.splint.rules.style.redundant-let
    noahtheduke.splint.rules.style.redundant-nested-call